+++ src/main/java/com/progex/tracker/item/service/ItemService.java
@@ -1,7 +1,7 @@
-package com.progex.tracker.service;
+package com.progex.tracker.item.service;
 
-import com.progex.tracker.entity.Category;
-import com.progex.tracker.entity.Item;
+import com.progex.tracker.category.entity.Category;
+import com.progex.tracker.item.entity.Item;
 
 import java.util.Optional;
