+++ src/main/java/com/progex/tracker/item/service/impl/ItemServiceImpl.java
@@ -1,10 +1,10 @@
-package com.progex.tracker.impl;
+package com.progex.tracker.item.service.impl;
 
-import com.progex.tracker.entity.Category;
-import com.progex.tracker.entity.Item;
-import com.progex.tracker.repo.ItemRepository;
-import com.progex.tracker.service.CategoryService;
-import com.progex.tracker.service.ItemService;
+import com.progex.tracker.category.entity.Category;
+import com.progex.tracker.item.entity.Item;
+import com.progex.tracker.item.repo.ItemRepository;
+import com.progex.tracker.item.service.ItemService;
+import com.progex.tracker.category.service.CategoryService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
