+++ src/test/java/com/progex/tracker/category/service/impl/CategoryServiceImplTest.java
@@ -1,12 +1,14 @@
-package com.progex.tracker.impl;
+package com.progex.tracker.category.service.impl;
 
-import com.progex.tracker.entity.Category;
-import com.progex.tracker.entity.Item;
-import com.progex.tracker.repo.CategoryRepository;
-import com.progex.tracker.repo.ItemRepository;
-import com.progex.tracker.service.CategoryService;
-import com.progex.tracker.service.ItemService;
+import com.progex.tracker.category.entity.Category;
+import com.progex.tracker.category.repo.CategoryRepository;
+import com.progex.tracker.category.service.CategoryService;
 import com.progex.tracker.exceptions.EntityNotFoundException;
+import com.progex.tracker.item.entity.Item;
+import com.progex.tracker.item.repo.ItemRepository;
+import com.progex.tracker.item.service.ItemService;
+import com.progex.tracker.item.service.impl.ItemServiceImpl;
+import com.progex.tracker.utility.TestUtils;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,7 +21,6 @@ import java.util.Optional;
 
 import static junit.framework.TestCase.assertTrue;
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.*;
 
@@ -70,9 +71,7 @@ public class CategoryServiceImplTest {
 
     @Test
     public void shouldSaveItemWhenAddingItem() {
-        Item item = new Item();
-        item.setName("Rice");
-        item.setDescription("Sri Lankan Rice");
+        Item item = TestUtils.getMockItem();
 
         Category category = new Category();
         category.setId(1);
